switched double amount to bigdecimal

+++ Bank Fund Domain/src/main/java/com/exalt_company/fund_domain/api/resource/Deposit.java
@@ -1,9 +1,10 @@
 package com.exalt_company.fund_domain.api.resource;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class Deposit extends FundResource {
-    public Deposit(UUID fundId, UUID fundOwnerId, double amount) {
+    public Deposit(UUID fundId, UUID fundOwnerId, BigDecimal amount) {
         super(fundId, fundOwnerId, amount);
     }
 }
